build: compile bench C shims with _GNU_SOURCE
The nightly `make rust-tests` has been failing on every platform since the
TopK numeric micro benchmarks landed: `optimizer_shim.c` reaches
`aggregate/reducer.h`, whose `FormatAggAlias` calls `rm_asprintf`. Shims are
compiled without `REDIS_MODULE_TARGET`, so `deps/rmalloc/rmalloc.h` takes its
non-module fallback and expands that to bare `asprintf` — which glibc only
declares under `_GNU_SOURCE`, and the root CMakeLists promotes implicit
declarations to errors. `ffi/build.rs` passes `-D_GNU_SOURCE` for exactly this
reason; the shims did not.

Centralise shim compilation in `build_utils::compile_c_bench_shim` so the
define, include roots and warning suppressions live in one place. Both bench
crates carried drifted copies of this setup; the vector one shares the missing
define and only escapes the error because `hybrid_shim.c` does not reach
`reducer.h`.

`REDIS_MODULE_TARGET` is deliberately left unset: CMake defines it only under
`USE_REDIS_ALLOCATOR`, so forcing it here could disagree with how the linked
`libredisearch_c_bundle.a` was built, and no shim calls `rm_*` anyway.

The include roots shared with the bindgen path now come from one
`c_include_roots`, with the shim-only roots appended so they cannot shadow a
shared one. `deps/RedisModulesSDK` is dropped from that list: the directory
does not exist in the tree, so it never resolved a header — `redismodule.h`
comes from `src/`. Both shim objects are byte-identical to those built from the
previous per-crate flags.
