+++ src/redisearch_rs/build_utils/src/lib.rs
@@ -61,6 +61,87 @@ pub fn rerun_if_c_changes(dir: &Path) -> std::io::Result<()> {
     rerun_if_changes(dir, &["c", "h"])
 }
 
+/// Include roots every Rust-side consumer of the RediSearch C headers needs.
+///
+/// Ordering is load-bearing: `src` must precede `src/redisearch_rs/headers`,
+/// which holds generated counterparts of `rlookup.h` and `ttl_table.h`. Swapping
+/// the two silently binds against the generated header instead of the C original.
+/// They are the only colliding header basenames across these roots.
+fn c_include_roots(root: &Path) -> Vec<PathBuf> {
+    vec![
+        root.join("src"),
+        root.join("deps"),
+        root.join("src").join("redisearch_rs").join("headers"),
+        root.join("deps").join("VectorSimilarity").join("src"),
+        root.join("src").join("buffer"),
+        root.join("src").join("ttl_table"),
+    ]
+}
+
+/// Compile a benchmark shim written in C against the RediSearch headers.
+///
+/// Bench crates use a small C shim to drive a production C iterator as the
+/// baseline for its Rust counterpart. Every such shim needs the same include
+/// roots, the same preprocessor defines, and the same warning suppressions, so
+/// they are centralised here: a shim that configures its own [`cc::Build`] from
+/// scratch silently drifts from how CMake compiles the very headers it includes.
+///
+/// `shim` is the path to the C file, relative to the calling crate's manifest
+/// directory. The compiled archive is named after the shim's file stem, and a
+/// `rerun-if-changed` is emitted for it.
+///
+/// # Defines
+///
+/// `_GNU_SOURCE` is required, for the same reason the bindgen path in
+/// `ffi/build.rs` passes it: shims are compiled without `REDIS_MODULE_TARGET`, so
+/// `deps/rmalloc/rmalloc.h` takes its non-module fallback and expands
+/// `rm_asprintf` to bare `asprintf` — which glibc's `<stdio.h>` only declares
+/// under `_GNU_SOURCE`, and the top-level `CMakeLists.txt` promotes implicit
+/// declarations to errors.
+///
+/// `REDIS_MODULE_TARGET` is deliberately *not* set. It would route `rm_malloc`
+/// through `RedisModule_Alloc`, but CMake defines it only when
+/// `USE_REDIS_ALLOCATOR` is on, so forcing it here could disagree with how
+/// `libredisearch_c_bundle.a` was actually built. No shim calls `rm_*` today —
+/// they use `RedisModule_Alloc` directly where the module owns the buffer.
+///
+/// # Panics
+///
+/// Panics if the repository root cannot be located, or if `shim` has no file stem.
+pub fn compile_c_bench_shim(shim: &str) {
+    let root = repository_root().expect("Could not find repository root");
+
+    // Beyond the shared roots, shims reach the iterator API, `rmalloc.h` and the
+    // trie headers directly. Appended, so they cannot shadow a shared root.
+    let mut includes = c_include_roots(&root);
+    includes.extend([
+        root.join("src").join("iterators"),
+        root.join("deps").join("rmalloc"),
+        root.join("src").join("trie"),
+    ]);
+
+    let lib_name = Path::new(shim)
+        .file_stem()
+        .and_then(|stem| stem.to_str())
+        .unwrap_or_else(|| panic!("bench shim path has no file stem: {shim}"));
+
+    cc::Build::new()
+        .file(shim)
+        .define("_GNU_SOURCE", None)
+        // Silence warnings originating from transitively-included RediSearch
+        // headers (static helpers, sign-compare in inline funcs, etc.) — they
+        // are not actionable from a shim and the main CMake build already
+        // suppresses them.
+        .flag_if_supported("-Wno-unused-function")
+        .flag_if_supported("-Wno-unused-variable")
+        .flag_if_supported("-Wno-unused-parameter")
+        .flag_if_supported("-Wno-sign-compare")
+        .includes(includes)
+        .compile(lib_name);
+
+    println!("cargo:rerun-if-changed={shim}");
+}
+
 /// Link all the relevant C dependencies to allow Rust (testing and benchmarking) code to invoke
 /// RediSearch C symbols.
 ///
@@ -218,15 +299,7 @@ pub fn generate_c_bindings(
     let root =
         repository_root().expect("Could not find repository root for static library linking");
 
-    let includes = vec![
-        root.join("deps").join("RedisModulesSDK"),
-        root.join("src"),
-        root.join("deps"),
-        root.join("src").join("redisearch_rs").join("headers"),
-        root.join("deps").join("VectorSimilarity").join("src"),
-        root.join("src").join("buffer"),
-        root.join("src").join("ttl_table"),
-    ];
+    let includes = c_include_roots(&root);
 
     let headers = headers
         .into_iter()
